Added way to create file container, also cleaned up things
