+++ constants.js
@@ -1,11 +1,9 @@
 'use strict';
 
 const BIG_FILE = 'File contents are too large. Limit is 256 MB.';
-const BINARY_FILE_ONLY = 'Files saved must be in binary format.';
 
 module.exports = {
 
-  BIG_FILE_MS,
-  BINARY_FILE_ONLY
-
-}'
+  BIG_FILE,
+  
+};
