+++ resources/push.js
@@ -7,6 +7,7 @@ const getFileDetailsFromRequest = utils.getFileDetailsFromRequest;
 
 // create a push
 const createPush = (z, bundle) => {
+
   const promise = z.request({
 
         url: `https://${bundle.inputData.platform}/v1/organizations/${bundle.inputData.orgId}/sources/${bundle.inputData.sourceId}/documents`,
